+++ src/components/ProgrammeList/ProgrammeList.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import AnalyticsDecorator from "../../storybook-decorators/AnalyticsDecorator";
-import tierListingFixture from "../../node-lib/curriculum-api/fixtures/tierListing.fixture";
+import { programmeListingFixture } from "../../node-lib/curriculum-api/fixtures/tierListing.fixture";
 
 import Component from ".";
 
@@ -15,8 +15,8 @@ const Template: ComponentStory<typeof Component> = (args) => (
   <Component {...args} />
 );
 
-export const TierList = Template.bind({});
+export const ProgrammeList = Template.bind({});
 
-TierList.args = {
-  ...tierListingFixture(),
+ProgrammeList.args = {
+  ...programmeListingFixture(),
 };