+++ src/components/ProgrammeList/ProgrammeListItem.test.tsx
@@ -3,7 +3,7 @@ import userEvent from "@testing-library/user-event";
 
 import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
 
-import TierListItem from "./TierListItem";
+import ProgrammeListItem from "./ProgrammeListItem";
 
 const tierSelected = jest.fn();
 jest.mock("../../context/Analytics/useAnalytics", () => ({
@@ -15,24 +15,28 @@ jest.mock("../../context/Analytics/useAnalytics", () => ({
   }),
 }));
 
-describe("TierListItem", () => {
+describe("ProgrammeListItem", () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
-  it("renders TierListItem", () => {
+  it("renders ProgrammeListItem", () => {
     renderWithTheme(
-      <TierListItem
+      <ProgrammeListItem
         subjectSlug="maths"
         subjectTitle="Maths"
         keyStageSlug="ks4"
         keyStageTitle="Key stage 4"
         tierTitle="Higher"
         tierSlug="higher"
-        totalUnitCount={3}
-        activeLessonCount={4}
+        unitCount={3}
+        lessonCount={4}
         programmeSlug="maths-secondary-ks4-higher"
         background="grey2"
+        tierDisplayOrder={null}
+        examBoardSlug={null}
+        examBoardTitle={null}
+        examBoardDisplayOrder={null}
       />
     );
 
@@ -41,17 +45,21 @@ describe("TierListItem", () => {
 
   it("calls tracking.tierSelected once, with correct props", async () => {
     renderWithTheme(
-      <TierListItem
+      <ProgrammeListItem
         subjectSlug="maths"
         subjectTitle="Maths"
         keyStageSlug="ks4"
         keyStageTitle="Key Stage 4"
         tierTitle="Higher"
         tierSlug="higher"
-        totalUnitCount={3}
-        activeLessonCount={4}
+        unitCount={3}
+        lessonCount={4}
         programmeSlug="maths-secondary-ks4-higher"
         background="grey2"
+        tierDisplayOrder={null}
+        examBoardSlug={null}
+        examBoardTitle={null}
+        examBoardDisplayOrder={null}
       />
     );