+++ src/node-lib/curriculum-api/index.ts
@@ -6,6 +6,12 @@ import config from "../../config/server";
 import OakError from "../../errors/OakError";
 import lessonListingSchema from "../curriculum-api-2023/queries/lessonListing/lessonListing.schema";
 import lessonDownloadsSchema from "../curriculum-api-2023/queries/downloads/downloads.schema";
+import { programmeListingSchema } from "../curriculum-api-2023/queries/programmeListing/programmeListing.schema";
+import {
+  baseLessonOverviewData,
+  lessonOverviewQuizData,
+  lessonQuizInfoData,
+} from "../curriculum-api-2023/queries/lessonOverview/lessonOverview.schema";
 
 import { getSdk } from "./generated/sdk";
 
@@ -21,7 +27,7 @@ const curriculumApiAuthKey = config.get("curriculumApiAuthKey");
  * cloud authentication (according to Thomas) but locally it's not working so
  * need to use admin-secret instead.
  */
-const curruclumApiAdminSecret = process.env.CURRICULUM_API_ADMIN_SECRET;
+const curriculumApiAdminSecret = process.env.CURRICULUM_API_ADMIN_SECRET;
 
 /**
  * TS complaining when Headers in not typed.
@@ -29,9 +35,9 @@ const curruclumApiAdminSecret = process.env.CURRICULUM_API_ADMIN_SECRET;
 type Headers =
   | { "x-hasura-admin-secret": string }
   | { "x-oak-auth-type": string; "x-oak-auth-key": string };
-const headers: Headers = curruclumApiAdminSecret
+const headers: Headers = curriculumApiAdminSecret
   ? {
-      "x-hasura-admin-secret": curruclumApiAdminSecret,
+      "x-hasura-admin-secret": curriculumApiAdminSecret,
     }
   : {
       "x-oak-auth-type": curriculumApiAuthType,
@@ -88,7 +94,6 @@ const unitsData = z.array(
     lessonCount: z.number().nullable(),
     quizCount: z.number().nullable(),
     unitStudyOrder: z.number(),
-    year: z.string(),
     expired: z.boolean().nullable(),
     expiredLessonCount: z.number().nullable(),
   })
@@ -99,8 +104,8 @@ const tiersData = z.array(
     tierSlug: z.string(),
     tierTitle: z.string(),
     tierProgrammeSlug: z.string(),
-    unitCount: z.number().nullable(),
-    lessonCount: z.number().nullable(),
+    unitCount: z.number().nullable().optional(),
+    lessonCount: z.number().nullable().optional(),
   })
 );
 
@@ -139,52 +144,6 @@ const lessonListingPaths = z.object({
   ),
 });
 
-const lessonOverviewQuizData = z.array(
-  z.object({
-    keyStageSlug: z.string(),
-    keyStageTitle: z.string(),
-    subjectSlug: z.string(),
-    subjectTitle: z.string(),
-    unitSlug: z.string(),
-    unitTitle: z.string(),
-    order: z.number().nullable().optional(),
-    title: z.string().nullable().optional(),
-    points: z.number().nullable().optional(),
-    required: z.boolean().nullable(),
-    choices: z.array(
-      z.object({
-        choice: z.string(),
-        image: z.string().nullable(),
-      })
-    ),
-    active: z.boolean(),
-    answer: z.union([z.array(z.string()), z.string()]),
-    type: z.string(),
-    quizType: z.string(),
-    images: z
-      .array(
-        z.union([
-          z.object({
-            title: z.string().nullable(),
-            images: z.array(z.string()),
-          }),
-          z.string(),
-        ])
-      )
-      .nullable(),
-    feedbackCorrect: z.string().nullable(),
-    feedbackIncorrect: z.string().nullable(),
-    displayNumber: z.string().nullable(),
-  })
-);
-
-const lessonQuizInfoData = z
-  .object({
-    title: z.string(),
-    questionCount: z.number(),
-  })
-  .nullable();
-
 const lessonOverviewPaths = z.object({
   lessons: z.array(
     z.object({
@@ -205,28 +164,7 @@ const lessonDownloadPaths = z.object({
   ),
 });
 
-const lessonOverviewData = z.object({
-  lessonSlug: z.string(),
-  lessonTitle: z.string(),
-  programmeSlug: z.string(),
-  unitTitle: z.string(),
-  unitSlug: z.string(),
-  keyStageSlug: z.string(),
-  keyStageTitle: z.string(),
-  subjectSlug: z.string(),
-  subjectTitle: z.string(),
-  coreContent: z.array(z.string().nullable()),
-  contentGuidance: z.string().nullable(),
-  equipmentRequired: z.string().nullable(),
-  presentationUrl: z.string().nullable(),
-  supervisionLevel: z.string().nullable(),
-  worksheetUrl: z.string().nullable(),
-  isWorksheetLandscape: z.boolean(),
-  hasCopyrightMaterial: z.boolean(),
-  videoMuxPlaybackId: z.string().nullable(),
-  videoWithSignLanguageMuxPlaybackId: z.string().nullable(),
-  transcriptSentences: z.array(z.string()).nullable(),
-  hasDownloadableResources: z.boolean().nullable(),
+export const lessonOverviewData = baseLessonOverviewData.extend({
   introQuiz: lessonOverviewQuizData,
   exitQuiz: lessonOverviewQuizData,
   introQuizInfo: lessonQuizInfoData,
@@ -234,7 +172,7 @@ const lessonOverviewData = z.object({
   expired: z.boolean(),
 });
 
-const programmesData = z.object({
+export const programmesData = z.object({
   subjectSlug: z.string(),
   subjectTitle: z.string(),
   keyStageSlug: z.string(),
@@ -247,15 +185,17 @@ const programmesData = z.object({
   programmeSlug: z.string(),
   tierSlug: z.string().nullable(),
   tierTitle: z.string().nullable().optional(),
+  lessonCount: z.number().optional(),
+  unitCount: z.number().optional(),
 });
 
 const programmesArray = z.array(programmesData);
 
-const subjectListingData = z.object({
+export const subjectListingData = z.object({
   keyStageSlug: z.string(),
   keyStageTitle: z.string(),
-  programmesAvailable: z.array(programmesData),
-  programmesUnavailable: z.array(programmesData),
+  subjects: z.array(programmesData),
+  subjectsUnavailable: z.array(programmesData),
 });
 
 const unitListingPaths = z.object({
@@ -270,6 +210,8 @@ const unitListingData = z.object({
   programmeSlug: z.string(),
   keyStageSlug: z.string(),
   keyStageTitle: z.string(),
+  examBoardSlug: z.string().nullable(),
+  examBoardTitle: z.string().nullable(),
   subjectSlug: z.string(),
   subjectTitle: z.string(),
   tierSlug: z.string().nullable(),
@@ -278,8 +220,8 @@ const unitListingData = z.object({
   units: unitsData,
   learningThemes: z.array(
     z.object({
-      learningThemeTitle: z.string().nullable(),
-      learningThemeSlug: z.string().nullable(),
+      themeTitle: z.string().nullable(),
+      themeSlug: z.string().nullable(),
     })
   ),
 });
@@ -409,11 +351,30 @@ const curriculumApi = {
         programmesArray.parse(programmesUnavailable)
       );
 
+    const addCurriculum2023Counts = (
+      programmes: ProgrammesData[] | undefined
+    ) => {
+      return programmes
+        ? programmes.map((programme) => {
+            return {
+              ...programme,
+              lessonCount: programme.nonDuplicateSubjectLessonCount,
+              unitCount: programme.nonDuplicateSubjectUnitCount,
+            };
+          })
+        : [];
+    };
+
     return subjectListingData.parse({
       keyStageSlug: keyStage.slug,
       keyStageTitle: keyStage.title,
-      programmesAvailable,
-      programmesUnavailable: filteredUnavailableProgrammeDuplicate || [],
+      subjects:
+        addCurriculum2023Counts(programmesArray.parse(programmesAvailable)) ||
+        [],
+      subjectsUnavailable:
+        addCurriculum2023Counts(
+          programmesArray.parse(filteredUnavailableProgrammeDuplicate)
+        ) || [],
     });
   },
   unitListingPaths: async () => {
@@ -429,8 +390,8 @@ const curriculumApi = {
 
     const programme = getFirstResultOrWarnOrFail()({ results: programmes });
     const learningThemes = units.map((unitWithTheme) => ({
-      learningThemeSlug: unitWithTheme?.themeSlug,
-      learningThemeTitle: unitWithTheme?.themeTitle || "No theme",
+      themeSlug: unitWithTheme?.themeSlug,
+      themeTitle: unitWithTheme?.themeTitle || "No theme",
     }));
 
     // !Refactor index signature to be more specific
@@ -440,10 +401,10 @@ const curriculumApi = {
         learningThemes.map((theme) => [JSON.stringify(theme), theme])
       ).values(),
     ].sort((a, b) => {
-      if (a.learningThemeTitle < b.learningThemeTitle) {
+      if (a.themeTitle < b.themeTitle) {
         return -1;
       }
-      if (a.learningThemeTitle > b.learningThemeTitle) {
+      if (a.themeTitle > b.themeTitle) {
         return 1;
       }
       return 0;
@@ -453,6 +414,8 @@ const curriculumApi = {
       programmeSlug: programme?.programmeSlug,
       keyStageSlug: programme?.keyStageSlug,
       keyStageTitle: programme?.keyStageTitle,
+      examBoardSlug: null,
+      examBoardTitle: null,
       subjectSlug: programme?.subjectSlug,
       subjectTitle: programme?.subjectTitle,
       totalUnitCount: programme?.totalUnitCount,
@@ -535,7 +498,36 @@ const curriculumApi = {
     const res = await sdk.tierListing(...args);
     const { programmes = [] } = transformMVCase(res);
 
-    return tierListingData.parse({ programmes });
+    const tierListingToProgrammeListing2013 = tierListingData
+      .parse({ programmes })
+      .programmes.map((programme) => {
+        return {
+          programmes: {
+            subjectSlug: programme.subjectSlug,
+            subjectTitle: programme.subjectTitle,
+            keyStageSlug: programme.keyStageSlug,
+            keyStageTitle: programme.keyStageTitle,
+            programmes: programmes.map((programme) => {
+              return {
+                programmeSlug: programme.programmeSlug,
+                subjectTitle: programme.subjectTitle,
+                unitCount: programme.totalUnitCount,
+                lessonCount: programme.activeLessonCount,
+                tierSlug: programme.tierSlug,
+                tierTitle: programme.tierTitle,
+                tierDisplayOrder: null,
+                examBoardSlug: null,
+                examBoardTitle: null,
+                examBoardDisplayOrder: null,
+              };
+            }),
+          },
+        };
+      });
+
+    return programmeListingSchema.parse(
+      tierListingToProgrammeListing2013[0]?.programmes
+    );
   },
 };