+++ src/pages/webinars/[webinarSlug].tsx
@@ -14,7 +14,6 @@ import { TeamMemberPreview, Webinar } from "../../common-lib/cms-types";
 import { getBlogWebinarPostBreadcrumbs } from "../../components/Breadcrumbs/getBreadcrumbs";
 import Box from "../../components/Box";
 import {
-  decorateWithIsr,
   getFallbackBlockingConfig,
   shouldSkipInitialBuild,
 } from "../../node-lib/isr";
@@ -25,6 +24,7 @@ import { BlogJsonLd } from "../../browser-lib/seo/getJsonLd";
 import { getVideoThumbnail } from "../../components/VideoPlayer/getVideoThumbnail";
 import useAnalytics from "../../context/Analytics/useAnalytics";
 import PostSingleLayout from "../../components/Posts/PostSingleLayout";
+import getPageProps from "../../node-lib/getPageProps";
 
 export type SerializedWebinar = Omit<Webinar, "date"> & {
   date: string;
@@ -104,40 +104,45 @@ export const getStaticProps: GetStaticProps<
   WebinarSinglePageProps,
   URLParams
 > = async (context) => {
-  const webinarSlug = context.params?.webinarSlug as string;
-  const isPreviewMode = context.preview === true;
+  return getPageProps({
+    page: "webinar-single::getStaticProps",
+    context,
+    getProps: async () => {
+      const webinarSlug = context.params?.webinarSlug as string;
+      const isPreviewMode = context.preview === true;
 
-  const webinarResult = await CMSClient.webinarBySlug(webinarSlug, {
-    previewMode: isPreviewMode,
-  });
+      const webinarResult = await CMSClient.webinarBySlug(webinarSlug, {
+        previewMode: isPreviewMode,
+      });
 
-  if (!webinarResult) {
-    return {
-      notFound: true,
-    };
-  }
+      if (!webinarResult) {
+        return {
+          notFound: true,
+        };
+      }
 
-  const webinarResults = await CMSClient.webinars();
+      const webinarResults = await CMSClient.webinars();
 
-  const categories = uniqBy(
-    "title",
-    webinarResults.map((w) => w.category)
-  ).sort((a, b) => (a.title < b.title ? -1 : 1));
+      const categories = uniqBy(
+        "title",
+        webinarResults.map((w) => w.category)
+      ).sort((a, b) => (a.title < b.title ? -1 : 1));
 
-  const webinar: SerializedWebinar = {
-    ...webinarResult,
-    date: webinarResult.date.toISOString(),
-    author: webinarResult.hosts[0], // make the first host equivalent to a blog author
-  };
+      const webinar: SerializedWebinar = {
+        ...webinarResult,
+        date: webinarResult.date.toISOString(),
+        author: webinarResult.hosts[0], // make the first host equivalent to a blog author
+      };
 
-  const results: GetStaticPropsResult<WebinarSinglePageProps> = {
-    props: {
-      webinar,
-      categories,
+      const results: GetStaticPropsResult<WebinarSinglePageProps> = {
+        props: {
+          webinar,
+          categories,
+        },
+      };
+      return results;
     },
-  };
-  const resultsWithIsr = decorateWithIsr(results);
-  return resultsWithIsr;
+  });
 };
 
 export default WebinarSinglePage;