+++ CHANGE_LOG.md
@@ -1,3 +1,108 @@
+# [1.202.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.201.2...v1.202.0) (2023-07-06)
+
+
+### Bug Fixes
+
+* **unit-listing:** fix test breaks ([5a1d9d8](https://github.com/oaknational/Oak-Web-Application/commit/5a1d9d8dfe58b0918796414456a75cc24152bea3))
+* **unitlisting:** increase test coverage for examboard presence ([5c03ac1](https://github.com/oaknational/Oak-Web-Application/commit/5c03ac18731281cf2338f332b405db772910e7e8))
+* **unitlisting:** repoint query for fix exam title ([164004c](https://github.com/oaknational/Oak-Web-Application/commit/164004cb668f2a18aaaf969aa81b8a3dd787d48c))
+
+
+### Features
+
+* **unit-listing:** add exam board to title if there is one ([06879cd](https://github.com/oaknational/Oak-Web-Application/commit/06879cd3e8a73cc1fad4859b512c52e5e4034bbc))
+* **unit-listing:** add examboard to the title is exists ([31d0d19](https://github.com/oaknational/Oak-Web-Application/commit/31d0d199d60851cad4be2252541f9c17638859c8))
+* **unit-listing:** prettier fix ([b27b4f8](https://github.com/oaknational/Oak-Web-Application/commit/b27b4f85cbc00dff88592610f9917e8b7e46cfac))
+
+## [1.201.2](https://github.com/oaknational/Oak-Web-Application/compare/v1.201.1...v1.201.2) (2023-07-04)
+
+
+### Bug Fixes
+
+* **lesson-listing:** connect lesson listing to mv to expose all units ([9e77a2d](https://github.com/oaknational/Oak-Web-Application/commit/9e77a2d44f0ba9f6bb21e8dcc16b6a89f752e081))
+
+## [1.201.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.201.0...v1.201.1) (2023-07-03)
+
+
+### Bug Fixes
+
+* remove log statements from components ([d9e48a6](https://github.com/oaknational/Oak-Web-Application/commit/d9e48a6f705a9933e001093d3a7a037db93af679))
+
+# [1.201.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.200.0...v1.201.0) (2023-07-03)
+
+
+### Bug Fixes
+
+* exam board tiers only come from exam board programmes ([9afeb65](https://github.com/oaknational/Oak-Web-Application/commit/9afeb659f3f98f8483c3462a1fe181246b4df27a))
+
+
+### Features
+
+* add exam boards to programme page ([8c04e14](https://github.com/oaknational/Oak-Web-Application/commit/8c04e1454d2489baa4b511ec8df220801985360f))
+
+# [1.200.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.199.0...v1.200.0) (2023-06-30)
+
+
+### Features
+
+* programme listing 2023 api ([6989837](https://github.com/oaknational/Oak-Web-Application/commit/698983746cffee184d12828dde1fe4df3db481d5))
+
+# [1.199.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.198.1...v1.199.0) (2023-06-29)
+
+
+### Features
+
+* **unit_lisitng:** connect unit listing page to the 2023 hasura instance ([a576f64](https://github.com/oaknational/Oak-Web-Application/commit/a576f644248779fa9042c6dd01d6b2639b791c19))
+* **unit_lisitng:** prettier fix ([9eb05d8](https://github.com/oaknational/Oak-Web-Application/commit/9eb05d860061b195a8bda1fa593087b270555d3b))
+* **unit_lisitng:** remove console logs ([cef2dc3](https://github.com/oaknational/Oak-Web-Application/commit/cef2dc3053af2f04e9b5a7937e41d8ea2d8cab5a))
+* **unit_lisitng:** remove console logs ([b504ab2](https://github.com/oaknational/Oak-Web-Application/commit/b504ab26d71dc30fa1e4873fc3aa7a9f25b7688b))
+* **unit_lisitng:** unit lisitng fix ([5d4f0a3](https://github.com/oaknational/Oak-Web-Application/commit/5d4f0a3cb3d14ca69653b0fe9319ef6f606f8ce2))
+
+## [1.198.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.198.0...v1.198.1) (2023-06-29)
+
+# [1.198.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.197.0...v1.198.0) (2023-06-28)
+
+
+### Bug Fixes
+
+* broken tests ([135a6da](https://github.com/oaknational/Oak-Web-Application/commit/135a6da559a14edb5c4ced9f569c2d6e3149d156))
+* lesson listing zod error test ([1e083fc](https://github.com/oaknational/Oak-Web-Application/commit/1e083fc6342d215a79a9f59dd2b1e7e99723e97b))
+* storybook ([4f35a15](https://github.com/oaknational/Oak-Web-Application/commit/4f35a151aa1f80f836cd8142ae3b41cd1deac01e))
+
+
+### Features
+
+* subject listing 2023 ([c323617](https://github.com/oaknational/Oak-Web-Application/commit/c3236177a939c97130c4e27718772456d2ee5022))
+
+# [1.197.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.196.1...v1.197.0) (2023-06-28)
+
+
+### Features
+
+* getPageProps() function with error handling and isr ([5d68f2f](https://github.com/oaknational/Oak-Web-Application/commit/5d68f2f22c15ad7578264d8beafd700d862dcf27))
+* isr error handling ([9b29ca6](https://github.com/oaknational/Oak-Web-Application/commit/9b29ca6bf8f50dfd2ec133cde378a0fc739b3ea3))
+
+## [1.196.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.196.0...v1.196.1) (2023-06-27)
+
+
+### Bug Fixes
+
+* change copy on newsletter form ([84c82e4](https://github.com/oaknational/Oak-Web-Application/commit/84c82e4a4d1fb8baee1a5374cae70b3b08582747))
+
+# [1.196.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.195.0...v1.196.0) (2023-06-27)
+
+
+### Bug Fixes
+
+* add missing npm ci step ([ced4c29](https://github.com/oaknational/Oak-Web-Application/commit/ced4c29fc749a21bcf382aaf513aa975452be898))
+* use forked version of updater action with explicit prNumber passed ([e59f7c0](https://github.com/oaknational/Oak-Web-Application/commit/e59f7c0fe6e05ee29c581e782c4edee17d6239a5))
+
+
+### Features
+
+* automatically add the deployment URL to PR descriptions ([763372a](https://github.com/oaknational/Oak-Web-Application/commit/763372a89f2850b1f27a6ddc589e124a7e5eb3b9))
+* swap {cloud link} for {owa_deployment_url} in PR template ([0dac2d9](https://github.com/oaknational/Oak-Web-Application/commit/0dac2d99d1c2967c7f7d778d737320105c6b8a9b))
+
 # [1.195.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.194.0...v1.195.0) (2023-06-26)